+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release Notes for Abandoned Cart
 
+## 1.5.0 - 2020-03-07
+### Added
+- New widget using Commerce 3 stats. Allows you to view recovered total based on date range.
+- Minor tweaks to dashboard stats and layout.
+
+### Changed
+- Plugin now required Commerce 3.0+ which also means Craft CMS 3.4+ is required.
+
 ## 1.4.1 - 2020-03-07
 ### Fixed
 - Fixed an issue relating to disabling 2nd reminder from `1.4.0`